Revamp install/uninstall
Including a fix so that we don't symlink a dir into itself!

+++ Rakefile
@@ -1,21 +1,20 @@
-require 'rake'
+require 'pathname'
 
-desc 'Hook our dotfiles into system-standard positions.'
+desc 'Hook dotfiles into system-standard positions.'
 task :install do
-  linkables = Dir.glob('**/*{.symlink}')
-
   skip_all = false
   overwrite_all = ENV['OVERWRITE_DOTFILES'] || false
   backup_all = false
 
-  linkables.each do |linkable|
+  Pathname.glob('**/*{.symlink}').each do |linkable|
+    skip = false
     overwrite = false
     backup = false
 
-    file = linkable.split('/').last.split('.symlink').last
-    target = %(#{ENV["HOME"]}/.#{file})
+    file = linkable.sub_ext('')
+    target = Pathname(ENV.fetch('HOME')).join(".#{file}")
 
-    if File.exists?(target) || File.symlink?(target)
+    if target.exist? || target.symlink?
       unless skip_all || overwrite_all || backup_all
         puts "File already exists: #{target}, what do you want to do? [s]kip, [S]kip all, [o]verwrite, [O]verwrite all, [b]ackup, [B]ackup all"
         case STDIN.gets.chomp
@@ -24,33 +23,34 @@ task :install do
         when 'O' then overwrite_all = true
         when 'B' then backup_all = true
         when 'S' then skip_all = true
-        when 's' then next
+        when 's' then skip = true
         end
       end
+
+      next if skip || skip_all
       FileUtils.rm_rf(target) if overwrite || overwrite_all
-      `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
+      `mv "#{target}" "#{target}.backup"` if backup || backup_all
     end
     `ln -s "$PWD/#{linkable}" "#{target}"`
   end
 end
 
 task :uninstall do
+  Pathname.glob('**/*{.symlink}').each do |linkable|
 
-  Dir.glob('**/*.symlink').each do |linkable|
-
-    file = linkable.split('/').last.split('.symlink').last
-    target = %(#{ENV["HOME"]}/.#{file})
+    file = linkable.sub_ext('')
+    target = Pathname(ENV.fetch('HOME')).join(".#{file}")
 
     # Remove all symlinks created during installation
-    if File.symlink?(target)
-      FileUtils.rm(target)
+    if target.symlink?
+      target.delete
     end
 
     # Replace any backups made during installation
-    if File.exists?(%(#{ENV["HOME"]}/.#{file}.backup))
-      `mv "$HOME/.#{file}.backup" "$HOME/.#{file}"`
+    backup = "#{target}.backup"
+    if backup.exist?
+      `mv "#{backup}" "#{target}"`
     end
-
   end
 end
